linter more refactoring
