+++ packages/generator-common/package.json
@@ -46,6 +46,7 @@
     "yargs": "^17.6.2"
   },
   "devDependencies": {
+    "@types/yargs": "^17.0.20",
     "mock-fs": "^5.2.0"
   }
 }
