+++ packages/generator-common/src/internal.ts
@@ -7,3 +7,4 @@ export * from './file-writer';
 export * from './package-description';
 export * from './ts-config';
 export * from './options-parser';
+export * from './cli-parser';
