+++ packages/generator/package.json
@@ -52,11 +52,9 @@
     "ts-morph": "^17.0.1",
     "typescript": "~4.9.4",
     "voca": "^1.4.0",
-    "winston": "^3.8.2",
-    "yargs": "^17.6.2"
+    "winston": "^3.8.2"
   },
   "devDependencies": {
-    "@types/yargs": "^17.0.20",
     "execa": "^5.0.0",
     "mock-fs": "^5.2.0",
     "prettier": "^2.8.3"
