+++ packages/openapi-generator/package.json
@@ -44,12 +44,10 @@
     "glob": "^8.1.0",
     "js-yaml": "^4.0.0",
     "openapi-types": "^12.1.0",
-    "swagger2openapi": "^7.0.4",
-    "yargs": "^17.6.2"
+    "swagger2openapi": "^7.0.4"
   },
   "devDependencies": {
     "@types/js-yaml": "^4.0.0",
-    "@types/yargs": "^17.0.20",
     "prettier": "^2.8.3",
     "mock-fs": "^5.2.0",
     "typescript": "~4.9.4"
